leave link nn
