+++ easylink/model/link_nn.py
@@ -1,0 +1,26 @@
+import torch
+import torch.nn.functional as F
+
+class LinkNN(torch.nn.Module):
+    def __init__(self, in_channels, hidden_channels, out_channels, num_layers, dropout):
+        super(LinkNN, self).__init__()
+        self.layers = torch.nn.ModuleList()
+        self.layers.append(torch.nn.Linear(in_channels, hidden_channels))
+        for _ in range(num_layers-2):
+            self.layers.append(torch.nn.Linear(
+                hidden_channels, hidden_channels))
+        self.layers.append(torch.nn.Linear(hidden_channels, out_channels))
+        self.dropout = dropout
+
+    def reset_parameters(self):
+        for layer in self.layers:
+            layer.reset_parameters()
+
+    def forward(self, x_i, x_j):
+        x = x_i * x_j
+        for layer in self.layers[:-1]:
+            x = layer(x)
+            x = F.relu(x)
+            x = F.dropout(x, p=self.dropout, training=self.training)
+        x = self.layers[-1](x)
+        return torch.sigmoid(x)
